# storepool: lock individual storepool details instead of all details