… pkg/kv/kvserver/allocator/allocatorimpl/allocator_test.go
… pkg/kv/kvserver/allocator/storepool/override_store_pool_test.go
… pkg/kv/kvserver/allocator/storepool/store_pool_test.go
… pkg/kv/kvserver/allocator_impl_test.go
… pkg/kv/kvserver/asim/gossip/gossip_test.go
… pkg/kv/kvserver/asim/state/state.go
… pkg/kv/kvserver/client_store_test.go
… pkg/kv/kvserver/helpers_test.go
… pkg/kv/kvserver/store_snapshot.go
… pkg/kv/kvserver/store_test.go