This commit takes mutex locks for individual storePool details instead of
the whole storePool map. This should reduce the mutex contention
especially at high store count clusters. Previously, every gossip store
update callback used to lock the whole StorePool.DetailsMu. This blocked
other accesses to any storepool.Detail.

Benchmark results:
```
name                                    old time/op    new time/op    delta
StorePoolReadsWithConcurrentUpdates-12    1.79ms ±40%    0.00ms ± 7%   -99.99%  (p=0.000 n=10+10)

name                                    old alloc/op   new alloc/op   delta
StorePoolReadsWithConcurrentUpdates-12    10.8kB ±57%     0.0kB ±33%   -99.99%  (p=0.000 n=10+10)

name                                    old allocs/op  new allocs/op  delta
StorePoolReadsWithConcurrentUpdates-12      44.2 ±22%       0.0       -100.00%  (p=0.000 n=9+10)
```

Note that the benchmark basically tests that updating a specific store detail
doesn't block accessing other store details.

References: #143145

Release note: None