@rafiss 2025-05-02
This mutex showed up in the [mutex profile](https://github.com/user-attachments/files/19170317/mutex.pb.gz) that I grabbed by running `./dev bench ./pkg/sql/tests --filter BenchmarkParallelSysbench --test-args '-test.mutexprofile=mutex.pb.gz  -test.cpu=8'` a few weeks ago.

I'm curious if we see an overall improvement on sysbench with this change.